Move addTaskToDefaultTasks and submitForm to logic.js
