+++ src/default.js
@@ -1,68 +1,7 @@
 import { createE } from './logic'
 import { Task } from './logic'
-
-const allTasks = {
-    default: [],
-    today: []
-};
-
-function addTaskToDefaultTasks(newTask, tasksObj, key) {
-    let defLocalStorage;
-    if (key === undefined) {
-        defLocalStorage = JSON.parse(localStorage.getItem('Default'));
-    } else {
-        defLocalStorage = JSON.parse(localStorage.getItem(JSON.stringify(key)));
-    }
-
-    if (defLocalStorage.length > 0) {
-        tasksObj.default = [];
-        defLocalStorage.map((e) => {
-            tasksObj.default.push(e);
-        });
-        tasksObj.default.push(newTask);
-        
-
-        if (key === undefined) {
-            localStorage.setItem("Default", JSON.stringify(allTasks.default));
-        } else {
-            localStorage.setItem(JSON.stringify(key), JSON.stringify(allTasks.default));
-        }
-
-    } else {
-        tasksObj.default.push(newTask);
-        if (key === undefined) {
-            localStorage.setItem("Default", JSON.stringify(allTasks.default));
-        } else {
-            localStorage.setItem(JSON.stringify(key), JSON.stringify(allTasks.default));
-        }
-    }
-}
-
-function submitForm(btn, key) {
-    btn.addEventListener('click', () => {
-        const title = document.querySelector('#title');
-        const description = document.querySelector('#description');
-        const dueDate = document.querySelector('#dueDate');
-        const priority = document.querySelector('#priority');
-        const notes = document.querySelector('#notes');
-        let checklist = document.querySelector('#checklist');
-        
-        if (checklist.checked) {
-            checklist.status = "Closed";
-        } else {
-            checklist.status = "Open";
-        }
-        if (title.value === '' || description.value === '' || dueDate.value === '' || priority.value === '' || notes.value === '') {
-            alert('Fields must be filled out'); // eslint-disable-line no-alert
-        } else {
-            const newTask = new Task(title.value, description.value, dueDate.value, priority.value, notes.value, checklist.status);
-            addTaskToDefaultTasks(newTask, allTasks, key);
-            let defPage = document.querySelector(".project-page");
-            defPage.appendChild(displayTable(key));
-        }
-    })
-    return btn;
-}
+import { addTaskToDefaultTasks } from './logic'
+import { submitForm } from './logic'
 
 function displayTaskForm(key) {
     const form = createE("div", false, "form");
@@ -246,4 +185,5 @@ function loadProject(key) {
     main.appendChild(createProject(key));
 }
 
+export { displayTable };
 export default loadProject
